Refactor import jobs
